+++ disp4/CPP/RigidElemFile.cpp
@@ -273,7 +273,7 @@ int CRigidElemFile::ReadBone( XMLIOBUF* xmliobuf )
 	else{
 		strcpy_s(bonename1, 256, bonename);
 		strcpy_s(bonename2, 256, bonename);
-		int headleng = jointnameptr - bonename;
+		int headleng = (int)(jointnameptr - bonename);
 		*(bonename2 + headleng) = 0;
 	}
 	CBone* curbone = m_model->GetBoneByName(bonename1);
@@ -335,7 +335,7 @@ int CRigidElemFile::ReadRE( XMLIOBUF* xmlbuf, CBone* curbone )
 	else{
 		strcpy_s(childname1, 256, childname);
 		strcpy_s(childname2, 256, childname);
-		int headleng = jointnameptr - childname;
+		int headleng = (int)(jointnameptr - childname);
 		*(childname2 + headleng) = 0;
 	}
 	CBone* chilbone = m_model->GetBoneByName(childname1);
